requirements.txt
google-genai
pydantic
pillow
ipykernel
python-dotenv
pydub